+++ src/main/java/com/example/demo/service/ItemService.java
@@ -57,7 +57,7 @@ public class ItemService {
 
 	// 商品を保存（新規追加）
 	@Transactional
-	public void saveNewItem(String name, String image, Integer price, String memo, Long userId, Integer categoryId) {
+	public void saveNewItem(String name, String image, Integer price, String memo, Long userId, Long categoryId) {
 		Item item = new Item();
 		item.setName(name);
 		item.setImage(image); // ここは "namekuzi.jpg" のようなファイル名だけ
